Update PaperMC download link
Have to update this since Paper's v1 api is being phased out and unfortunately there's no native alternative to download latest. This admittedly isn't much of a concern since 1.16.1 is a version no longer being updated, but it's definitely something to be wrestled with should this update to the latest version.